Get rid of the most of Eclipse warnings
git-svn-id: svn+ssh://svn.code.sf.net/p/cmusphinx/code/trunk/sphinx4@12318 94700074-3cef-4d97-a70e-9c8c206c02f5

+++ sphinx4-core/src/main/java/edu/cmu/sphinx/frontend/frequencywarp/LinearPredictor.java
@@ -149,18 +149,18 @@ public class LinearPredictor {
         cepstra[1] = -ARParameters[1];
 
         for (i = 2; i < Math.min(cepstrumOrder, order + 1); i++) {
-            sum = (double) i * ARParameters[i];
+            sum = i * ARParameters[i];
             for (int j = 1; j < i; j++) {
-                sum += ARParameters[j] * cepstra[i - j] * (double) (i - j);
+                sum += ARParameters[j] * cepstra[i - j] * (i - j);
             }
-            cepstra[i] = -sum / (double) i;
+            cepstra[i] = -sum / i;
         }
         for (; i < cepstrumOrder; i++) { // Only if cepstrumOrder > order+1
             sum = 0;
             for (int j = 1; j <= order; j++) {
-                sum += ARParameters[j] * cepstra[i - j] * (double) (i - j);
+                sum += ARParameters[j] * cepstra[i - j] * (i - j);
             }
-            cepstra[i] = -sum / (double) i;
+            cepstra[i] = -sum / i;
         }
         return cepstra;
     }
